chore: move inteface and styles folder outside of src
