+++ interface/static.ts
