+++ src/card/CardItem.tsx
@@ -1,5 +1,5 @@
 import React, { useState } from "react";
-import { CardType } from "../interface/static";
+import { CardType } from "../../interface/static";
 
 import styled from "styled-components";
 import LikeButton from "../../common/OrangeLikeButton";
