+++ src/home/WelcomeText.tsx
@@ -1,6 +1,6 @@
 import React from "react";
-import { WELCOME_MESSAGE_FIRST } from "../interface/static";
-import { WELCOME_MESSAGE_SECOND } from "../interface/static";
+import { WELCOME_MESSAGE_FIRST } from "../../interface/static";
+import { WELCOME_MESSAGE_SECOND } from "../../interface/static";
 import styled from "styled-components";
 
 interface WelcomeTextProps {
